# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "App AI Gateway API",
    "version": "0.1.0",
    "description": "A multi-tenant, provider-native LLM proxy for mobile applications and trusted server backends."
  },
  "servers": [
    {
      "url": "https://gateway.example.com",
      "description": "Replace with your deployed gateway origin"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "ManagementBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "agw_mgmt_…",
        "description": "An organization-scoped management API key created by an owner or admin."
      },
      "OperatorSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "agw_operator_auth.session_token",
        "description": "The Better Auth operator session cookie. Admin requests also send x-console-request: 1."
      },
      "GatewayBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "A gateway access token, or an application API key for an issuer-less API-key app."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "data": {
                "type": "object",
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "number"
                    },
                    {
                      "type": "boolean"
                    }
                  ]
                },
                "description": "Machine-readable facts about this rejection, present only where the code alone is not actionable. A billing_request_quota_exceeded rejection carries limit, used, and resetAt; an app_rate_limited or app_budget_exhausted rejection carries scope, either user or app."
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "AppAttestRegisterRequest": {
        "type": "object",
        "properties": {
          "issuer_token": {
            "type": "string",
            "minLength": 1
          },
          "key_id": {
            "type": "string",
            "minLength": 1
          },
          "attestation": {
            "type": "string",
            "minLength": 1
          },
          "challenge": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "issuer_token",
          "key_id",
          "attestation",
          "challenge"
        ]
      },
      "AppAttestTokenRequest": {
        "type": "object",
        "properties": {
          "issuer_token": {
            "type": "string",
            "minLength": 1
          },
          "key_id": {
            "type": "string",
            "minLength": 1
          },
          "assertion": {
            "type": "string",
            "minLength": 1
          },
          "challenge": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "issuer_token",
          "key_id",
          "assertion",
          "challenge"
        ]
      },
      "ApiKeyTokenRequest": {
        "type": "object",
        "properties": {
          "issuer_token": {
            "type": "string",
            "minLength": 1
          },
          "api_key": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "issuer_token",
          "api_key"
        ],
        "additionalProperties": false
      },
...
    },
...
  },
  "paths": {
...
    "/v1/apps/{app}/auth/challenge": {
      "post": {
        "tags": [
          "Application authentication"
        ],
        "operationId": "createAppAttestChallenge",
        "summary": "Create an App Attest challenge",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "my-app"
            },
            "required": true,
            "name": "app",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "A five-minute, single-use challenge.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "challenge": {
                      "type": "string"
                    },
                    "expires_in": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "challenge",
                    "expires_in"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "No billing plan resolves for the organization.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated identity is not allowed to perform this operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/apps/{app}/auth/register": {
      "post": {
        "tags": [
          "Application authentication"
        ],
        "operationId": "registerAppAttestKey",
        "summary": "Register an App Attest key",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "my-app"
            },
            "required": true,
            "name": "app",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppAttestRegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The key was registered for the verified issuer identity.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user_id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "user_id"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "No billing plan resolves for the organization.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "`issuer_token_rejected` — the token did not verify; get a fresh one, retry once, then fail. `issuer_claims_missing` — the token is valid but a required entitlement claim has not propagated yet; do not re-authenticate, wait and retry. `auth_required` — the key was refused or the user is blocked. `attest_failed` — the App Attest proof did not hold.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "`issuer_verification_unavailable` — the gateway could not reach or read the issuer's keys, so the token was never judged. Retry with backoff.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/apps/{app}/auth/token": {
      "post": {
        "tags": [
          "Application authentication"
        ],
        "operationId": "exchangeGatewayToken",
        "summary": "Exchange issuer identity and client proof for a gateway token",
        "description": "App Attest clients send key_id, assertion, and challenge. Issuer-backed API-key clients send api_key and issuer_token. The verified issuer claim is always the resulting user identity.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "my-app"
            },
            "required": true,
            "name": "app",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AppAttestTokenRequest"
                  },
                  {
                    "$ref": "#/components/schemas/ApiKeyTokenRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A short-lived gateway access token.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "expires_in": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "access_token",
                    "expires_in"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "No billing plan resolves for the organization.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "`issuer_token_rejected` — the token did not verify; get a fresh one, retry once, then fail. `issuer_claims_missing` — the token is valid but a required entitlement claim has not propagated yet; do not re-authenticate, wait and retry. `auth_required` — the key was refused or the user is blocked. `attest_failed` — the App Attest proof did not hold.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "`issuer_verification_unavailable` — the gateway could not reach or read the issuer's keys, so the token was never judged. Retry with backoff.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}